small changes to strlen and lstmap

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: mamendes <mamendes@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/09 18:15:31 by mamendes          #+#    #+#             */
-/*   Updated: 2026/04/27 19:55:15 by mamendes         ###   ########.fr       */
+/*   Updated: 2026/05/07 19:18:33 by mamendes         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
